Adding functionality and fixing bugs
Added functionality to click on a field in a task post and entering edit
mode if eligible.
Fixing 'data leak' with post, rest services were returning the whole
post object.
Fixing error with task not found. No post was assigned and an error was
thrown. Fixed by removing the leak from above.